refactor(tests): remove unnecessary imports, and change new Command class to use Command.with method
