+++ infrastructure/src/test/java/com/kaua/ecommerce/users/infrastructure/amqp/AccountCreatedListenerTest.java
@@ -1,11 +1,11 @@
 package com.kaua.ecommerce.users.infrastructure.amqp;
 
 import com.kaua.ecommerce.users.AmqpTest;
+import com.kaua.ecommerce.users.application.either.Either;
+import com.kaua.ecommerce.users.application.gateways.AccountGateway;
 import com.kaua.ecommerce.users.application.usecases.account.mail.create.CreateAccountMailCommand;
 import com.kaua.ecommerce.users.application.usecases.account.mail.create.CreateAccountMailOutput;
 import com.kaua.ecommerce.users.application.usecases.account.mail.create.CreateAccountMailUseCase;
-import com.kaua.ecommerce.users.application.either.Either;
-import com.kaua.ecommerce.users.application.gateways.AccountGateway;
 import com.kaua.ecommerce.users.domain.accounts.Account;
 import com.kaua.ecommerce.users.domain.accounts.AccountCreatedEvent;
 import com.kaua.ecommerce.users.domain.accounts.mail.AccountMailType;
